added actions

+++ src/main/java/com/javarush/island/uspenskaya/util/Randomizer.java
@@ -9,7 +9,7 @@ public class Randomizer {
     public static int getRnd (int from, int to){
         return ThreadLocalRandom.current().nextInt(from,to);
     }
-    public static int getRndDouble (double from, double to){
+    public static double getRndDouble (double from, double to){
         return ThreadLocalRandom.current().nextDouble(from,to);
     }
 }
